Verify AST in form tests as well
