+++ test/form/samples/supports-core-js/_config.js
@@ -1,7 +1,6 @@
 module.exports = defineTest({
-	// TODO SWC skipped for now as problematic utf-8 characters are encoded differently in the AST
-	skip: true,
 	description: 'supports core-js',
+	verifyAst: false,
 	options: {
 		// check against tree-shake: false when updating the polyfill
 		treeshake: true,
