+++ test/function/index.js
@@ -1,13 +1,16 @@
 const assert = require('node:assert');
 const path = require('node:path');
-const { Parser } = require('acorn');
-const { importAssertions } = require('acorn-import-assertions');
 /**
  * @type {import('../../src/rollup/types')} Rollup
  */
 // @ts-expect-error not included in types
 const rollup = require('../../dist/rollup');
-const { compareError, compareLogs, runTestSuiteWithSamples } = require('../utils.js');
+const {
+	compareError,
+	compareLogs,
+	runTestSuiteWithSamples,
+	verifyAstPlugin
+} = require('../utils.js');
 
 function requireWithContext(code, context, exports) {
 	const module = { exports };
@@ -218,34 +221,3 @@ runTestSuiteWithSamples(
 		);
 	}
 );
-
-const acornParser = Parser.extend(importAssertions);
-
-const verifyAstPlugin = {
-	name: 'verify-ast',
-	moduleParsed: ({ ast, code }) => {
-		const acornAst = acornParser.parse(code, { ecmaVersion: 'latest', sourceType: 'module' });
-		assert.deepStrictEqual(
-			JSON.parse(JSON.stringify(ast, replaceStringifyValues), reviveStringifyValues),
-			JSON.parse(JSON.stringify(acornAst, replaceStringifyValues), reviveStringifyValues)
-		);
-	}
-};
-
-const replaceStringifyValues = (key, value) =>
-	key.startsWith('_')
-		? undefined
-		: typeof value == 'bigint'
-		? `~BigInt${value.toString()}`
-		: value instanceof RegExp
-		? `~RegExp${JSON.stringify({ flags: value.flags, source: value.source })}`
-		: value;
-
-const reviveStringifyValues = (_, value) =>
-	typeof value === 'string'
-		? value.startsWith('~BigInt')
-			? BigInt(value.slice(7))
-			: value.startsWith('~RegExp')
-			? new RegExp(JSON.parse(value.slice(7)).source, JSON.parse(value.slice(7)).flags)
-			: value
-		: value;
